Moved method altHeaderMetadataToVCFHeaderLine() to VCFUtils
