+++ src/main/kotlin/net/maizegenetics/phgv2/utils/VCFUtils.kt
@@ -4,27 +4,30 @@ package net.maizegenetics.phgv2.utils
 
 import htsjdk.variant.vcf.VCFAltHeaderLine
 import htsjdk.variant.vcf.VCFHeader
+import htsjdk.variant.vcf.VCFHeaderVersion
 import org.apache.logging.log4j.LogManager
 
 private val myLogger = LogManager.getLogger("net.maizegenetics.phgv2.utils.VCFUtils")
 
 // Making Number a string as VCF allows for '.'
-data class AltHeaderMetaData(val id: String, val description:String, val source:String, val sampleName:String,
-                             val regions:List<Pair<Position, Position>>, val checksum:String, val refRange:String)
+data class AltHeaderMetaData(
+    val id: String, val description: String, val source: String, val sampleName: String,
+    val regions: List<Pair<Position, Position>>, val checksum: String, val refRange: String
+)
 
 /**
  * Helper function to parse out the ALT headers from the VCF file.
  *
  * We need to do a bit more involved parsing in this function as we cannot use the .getOtherHeaders() call from HTSJDK.
- * For some reason this only returns the first header when called and we need all of them.
- * The work around is that we can get all the metadata, filter out any that are not ALT then parse the ALT header using normal string parsing.
+ * For some reason this only returns the first header when called, and we need all of them.
+ * The workaround is that we can get all the metadata, filter out any that are not ALT then parse the ALT header using normal string parsing.
  * To make this easy, we just parse each piece of metadata into a key-value pair and then store in a map.
-*/
-fun parseALTHeader(header: VCFHeader) : Map<String, AltHeaderMetaData> {
+ */
+fun parseALTHeader(header: VCFHeader): Map<String, AltHeaderMetaData> {
 
-    return header.metaDataInInputOrder.filter { it.key =="ALT" }
-        .map{it as VCFAltHeaderLine }
-        .map{it.genericFields}
+    return header.metaDataInInputOrder.filter { it.key == "ALT" }
+        .map { it as VCFAltHeaderLine }
+        .map { it.genericFields }
         .associateBy { it["ID"]!! }
         .map {
             check(it.value.containsKey("ID")) { "ALT Header does not contain ID" }
@@ -51,10 +54,23 @@ fun parseALTHeader(header: VCFHeader) : Map<String, AltHeaderMetaData> {
 /**
  * Function to parse the regions from the ALT header.
  */
-fun parseRegions(regions: String) : List<Pair<Position, Position>> {
+fun parseRegions(regions: String): List<Pair<Position, Position>> {
     return regions.split(",").map { it.split(":") }.map {
         val positions = it[1].split("-").map { position -> position.toInt() }
         check(positions.size == 2) { "Region ${it} is not in the correct format.  It needs to be in the form: chr:stPos-endPos." }
-        Pair(Position(it[0],positions[0]), Position(it[0],positions[1]))
+        Pair(Position(it[0], positions[0]), Position(it[0], positions[1]))
     }
+}
+
+fun altHeaderMetadataToVCFHeaderLine(altHeaderData: AltHeaderMetaData): VCFAltHeaderLine {
+
+    return VCFAltHeaderLine(
+        "<ID=${altHeaderData.id}, " +
+                "Description=\"haplotype data for line: ${altHeaderData.sampleName}\">," +
+                "Source=\"${altHeaderData.source}\",SampleName=\"${altHeaderData.sampleName}\"," +
+                "Regions=\"${altHeaderData.regions.joinToString(",") { "${it.first.contig}:${it.first.position}-${it.second.position}" }}\"," +
+                "Checksum=\"Md5\",RefRange=\"${altHeaderData.refRange}\">",
+        VCFHeaderVersion.VCF4_2
+    )
+
 }
